Move useSDKMethod to its own file
